requirements.txt
litellm>=1.0.0
faiss-cpu
rank-bm25
numpy
pandas
scikit-learn
joblib
python-dotenv
ollama
